Inference for numerical data
In 2004, the state of North Carolina released a large dataset containing information on births recorded in this state. This dataset is useful to researchers studying the relationship between habits and practices of expectant mothers and the birth of their children. We will work with a random sample of observations from this data set.